remove old impl from occlusion_spot
