+++ planning/behavior_velocity_planner/autoware_behavior_velocity_occlusion_spot_module/plugins.xml
@@ -1,6 +1,4 @@
 <library path="autoware_behavior_velocity_occlusion_spot_module">
   <class type="autoware::behavior_velocity_planner::OcclusionSpotModulePlugin"
-    base_class_type="autoware::behavior_velocity_planner::PluginInterface" />
-  <class type="autoware::behavior_velocity_planner::experimental::OcclusionSpotModulePlugin"
     base_class_type="autoware::behavior_velocity_planner::experimental::PluginInterface" />
 </library>
