+++ planning/behavior_velocity_planner/autoware_behavior_velocity_occlusion_spot_module/test/src/test_occlusion_spot_utils.cpp
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 #include "gtest/gtest.h"
-#include "occlusion_spot_utils.hpp"
 #include "utils.hpp"
 
 #include <autoware/behavior_velocity_planner_common/utilization/path_utilization.hpp>
@@ -35,18 +34,15 @@ using autoware_planning_msgs::msg::PathPoint;
 
 TEST(calcSlowDownPointsForPossibleCollision, TooManyPossibleCollisions)
 {
-  using autoware::behavior_velocity_planner::occlusion_spot_utils::
-    calcSlowDownPointsForPossibleCollision;
-  using autoware::behavior_velocity_planner::occlusion_spot_utils::PossibleCollisionInfo;
-  using std::chrono::duration;
+  using autoware::behavior_velocity_planner::utils::calcSlowDownPointsForPossibleCollision;
+  using autoware::behavior_velocity_planner::utils::PossibleCollisionInfo;
   using std::chrono::duration_cast;
   using std::chrono::high_resolution_clock;
   using std::chrono::microseconds;
-  std::vector<PossibleCollisionInfo> possible_collisions;
-  size_t num = 2000;
+  std::vector<PossibleCollisionInfo> possible_collisions{};
+  const size_t num = 2000;
   // make a path with 2000 points from x=0 to x=4
-  autoware_internal_planning_msgs::msg::PathWithLaneId path =
-    test::generatePath(0.0, 3.0, 4.0, 3.0, num);
+  const auto path = test::generatePath(0.0, 3.0, 4.0, 3.0, num);
   // make 2000 possible collision from x=0 to x=10
   test::generatePossibleCollisions(possible_collisions, 0.0, 3.0, 4.0, 3.0, num);
 
@@ -58,22 +54,19 @@ TEST(calcSlowDownPointsForPossibleCollision, TooManyPossibleCollisions)
    */
 
   auto start_naive = high_resolution_clock::now();
-  calcSlowDownPointsForPossibleCollision(0, path, 0, possible_collisions);
+  calcSlowDownPointsForPossibleCollision(path, 0, possible_collisions);
 
   auto end_naive = high_resolution_clock::now();
   // 2000 path * 2000 possible collisions
-  EXPECT_EQ(possible_collisions.size(), size_t{num});
-  EXPECT_EQ(path.points.size(), size_t{num});
+  EXPECT_EQ(possible_collisions.size(), num);
   std::cout << " runtime (microsec) "
             << duration_cast<microseconds>(end_naive - start_naive).count() << std::endl;
 }
 
 TEST(calcSlowDownPointsForPossibleCollision, ConsiderSignedOffset)
 {
-  using autoware::behavior_velocity_planner::occlusion_spot_utils::
-    calcSlowDownPointsForPossibleCollision;
-  using autoware::behavior_velocity_planner::occlusion_spot_utils::PossibleCollisionInfo;
-  using std::chrono::duration;
+  using autoware::behavior_velocity_planner::utils::calcSlowDownPointsForPossibleCollision;
+  using autoware::behavior_velocity_planner::utils::PossibleCollisionInfo;
   using std::chrono::duration_cast;
   using std::chrono::high_resolution_clock;
   using std::chrono::microseconds;
@@ -81,10 +74,12 @@ TEST(calcSlowDownPointsForPossibleCollision, ConsiderSignedOffset)
   // for public
   {
     const double offset_from_start_to_ego = 0;
-    PathWithLaneId path = test::generatePath(0.0, 3.0, 6.0, 3.0, 7);
-    for (size_t i = 0; i < path.points.size(); i++) {
-      path.points[i].point.longitudinal_velocity_mps = static_cast<double>(i);
+    auto path = test::generatePath(0.0, 3.0, 6.0, 3.0, 7);
+    auto path_points = path.restore();
+    for (size_t i = 0; i < path_points.size(); i++) {
+      path_points[i].point.longitudinal_velocity_mps = static_cast<double>(i);
     }
+    path = autoware::experimental::trajectory::pretty_build(path_points).value();
     test::generatePossibleCollisions(pcs, 3.0, 3.0, 6.0, 3.0, 3);
     /**
      * @brief generated path and possible collisions : path start from 2 to 6
@@ -97,10 +92,10 @@ TEST(calcSlowDownPointsForPossibleCollision, ConsiderSignedOffset)
      *    p : path
      *    c : collision
      */
-    calcSlowDownPointsForPossibleCollision(0, path, -offset_from_start_to_ego, pcs);
+    calcSlowDownPointsForPossibleCollision(path, -offset_from_start_to_ego, pcs);
     if (pcs[0].collision_with_margin.longitudinal_velocity_mps != 3.0) {
-      for (size_t i = 0; i < path.points.size(); i++) {
-        std::cout << "v : " << path.points[i].point.longitudinal_velocity_mps << "\t";
+      for (size_t i = 0; i < path_points.size(); i++) {
+        std::cout << "v : " << path_points[i].point.longitudinal_velocity_mps << "\t";
       }
       std::cout << std::endl;
       for (const auto & pc : pcs) {
